% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_top_21}
\alias{calculate_top_21}
\title{Master function that calculates the top 21 players for each team
in a given week subject to positional constraints}
\usage{
calculate_top_21(
  season,
  league_id,
  api_key,
  current_week,
  force_update = FALSE
)
}
\arguments{
\item{season}{An NFL season}

\item{league_id}{An MFL league ID}

\item{api_key}{An MFL api key. See Help > Developer API while logged into
any of your leagues on myfantasyleague.com and scroll to the bottom of the page.}

\item{current_week}{The current NFL week. If the season is over, use 18.}

\item{force_update}{Defaults to FALSE. Set to TRUE if you are getting errors like "Couldn't resolve host name"}
}
\value{
A filtered tibble of roster data with the top 21 players on each team
for each week of play, given the roster constraints defined in solve_for_pp
}
\description{
Master function that calculates the top 21 players for each team
in a given week subject to positional constraints
}
