% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_base_url}
\alias{get_base_url}
\title{Gets a league's base url from the MFL API and stores it in a database.}
\usage{
get_base_url(league_id)
}
\arguments{
\item{league_id}{An MFL league identifier}
}
\value{
A tibble with the base URL string for the league and an updated date.
}
\description{
Gets a league's base url from the MFL API and stores it in a database.
}
