% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_player_per_game_scores}
\alias{get_player_per_game_scores}
\title{Title}
\usage{
get_player_per_game_scores(season, league_id, api_key, current_week)
}
\arguments{
\item{season}{An NFL season}

\item{league_id}{An MFL league identifier}

\item{api_key}{An MFL api key. See Help > Developer API while logged into}

\item{current_week}{The most current week for which data exists}
}
\value{
a tibble of player stats for each week
}
\description{
Title
}
