% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_players}
\alias{get_players}
\title{Get all player names in a league}
\usage{
get_players(season, league_id, api_key)
}
\arguments{
\item{season}{An NFL season}

\item{league_id}{An MFL league identifier}

\item{api_key}{An MFL api key. See Help > Developer API while logged into
any of your leagues on myfantasyleague.com and scroll to the bottom of the page.}
}
\value{
A tibble of player data including draft_team, name, position, team, draft_year,
stats_global_id, id, twitter_username, fleaflicker_id, stats_id, cbs_id, sportsdata_id,
rotowire_id, height, rotoworld_id, college, draft_round, weight, jersey, birthdate, draft_pick, espn_id,
nfl_id, and a binary flag for rookie
}
\description{
Get all player names in a league
}
\examples{
# api_key <- "your API key here"
# league_id <- 59406
# season <- 2023
# get_players(season, league_id, api_key)
}
