% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_teams}
\alias{get_teams}
\title{Get all teams from a given MFL league}
\usage{
get_teams(season, league_id, api_key)
}
\arguments{
\item{season}{An NFL season}

\item{league_id}{An MFL league identifier}

\item{api_key}{An MFL api key. See Help > Developer API while logged into
any of your leagues on myfantasyleague.com and scroll to the bottom of the page.}
}
\value{
a tibble of team data including team id, team name, division, available cap,
and conference
}
\description{
Get all teams from a given MFL league
}
