% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_week_roster}
\alias{get_week_roster}
\title{Get the rosters for each team for a given week. Includes players on Taxi and
IR.}
\usage{
get_week_roster(league_id, season, week, api_key)
}
\arguments{
\item{league_id}{An MFL league ID}

\item{season}{An NFL season}

\item{week}{A week in the NFL season}

\item{api_key}{An MFL api key. See Help > Developer API while logged into
any of your leagues on myfantasyleague.com and scroll to the bottom of the page.}
}
\value{
A tibble of rosters for each team in the league for a given week
}
\description{
Get the rosters for each team for a given week. Includes players on Taxi and
IR.
}
