% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_weekly_results}
\alias{get_weekly_results}
\title{Gets weekly scores for each team's players up to the current week.
Includes players on Taxi and IR.}
\usage{
get_weekly_results(season, league_id, api_key, current_week)
}
\arguments{
\item{season}{An NFL season}

\item{league_id}{An MFL league ID}

\item{api_key}{An MFL api key. See Help > Developer API while logged into
any of your leagues on myfantasyleague.com and scroll to the bottom of the page.}

\item{current_week}{The current NFL week. If the season is over, use 18.}
}
\value{
A tibble of player scores by team
}
\description{
Gets weekly scores for each team's players up to the current week.
Includes players on Taxi and IR.
}
