% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_base_url}
\alias{load_base_url}
\title{Loads the base URL for a league from a database or requests a updated URL from MFL.}
\usage{
load_base_url(league_id)
}
\arguments{
\item{league_id}{An MFL league identifier}
}
\value{
A tibble with the base URL string for the league and an updated date.
}
\description{
Loads the base URL for a league from a database or requests a updated URL from MFL.
}
