% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_nfc_pp_table}
\alias{make_nfc_pp_table}
\title{Make a table of lowest potential points scored by team, filtered to the NFC}
\usage{
make_nfc_pp_table(df)
}
\arguments{
\item{df}{A tibble returned from calculate_top_21}
}
\value{
a GT object. Use View() to send it to the RStudio viewer or save, print, etc.
}
\description{
Make a table of lowest potential points scored by team, filtered to the NFC
}
