% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{solve_for_pp}
\alias{solve_for_pp}
\title{Using linear programming, solves for the optimal 21 player lineup given
a series of constraints.}
\usage{
solve_for_pp(roster)
}
\arguments{
\item{roster}{A roster to optimize for.}
}
\value{
a tibble of 21 players
}
\description{
Using linear programming, solves for the optimal 21 player lineup given
a series of constraints.
}
